+++ addons/filamind_l10n_eu_iot_scale_cert/models/__init__.py
@@ -1,0 +1,2 @@
+from . import iot_device
+from . import scale_weighing_event